% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHARMa_ggplot.R
\name{DHARMa_ggplot}
\alias{DHARMa_ggplot}
\title{Function to make ggplot versions of DHARMa plots}
\usage{
DHARMa_ggplot(model)
}
\arguments{
\item{model}{A model which is compatible with the DHARMa package}
}
\value{
A ggplot object containing DHARMa equivalent plots
}
\description{
Function to make ggplot versions of DHARMa plots
}
\examples{
# plots <- DHARMa_ggplot(g3) # specify model
}
