% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_process_wind_data.R
\name{download_process_wind_data}
\alias{download_process_wind_data}
\title{Download and process wind data for specified locations and time periods}
\usage{
download_process_wind_data(
  locations,
  start_year = 2009,
  end_year = 2024,
  output_dir = "Data/Wind Data",
  save_interim = TRUE,
  download_u = TRUE,
  download_v = TRUE,
  overwrite_existing = FALSE
)
}
\arguments{
\item{locations}{Required dataframe with Lat, Lon, and optional location name}

\item{start_year}{Start year for data (default: 2009)}

\item{end_year}{End year for data (default: 2024)}

\item{output_dir}{Directory for output files}

\item{save_interim}{Save intermediate files (default: TRUE)}

\item{download_u}{Download U wind data (default T)}

\item{download_v}{Download U wind data  (default T)}

\item{overwrite_existing}{Overwrite existing files (default: FALSE)}
}
\value{
Dataframe containing daily wind speed and direction (going towards) for each location in the provided datafarme.
}
\description{
Download and process wind data for specified locations and time periods
}
\examples{
#Example usage:
sample_locations <- data.frame(
  Lat = c(-32.27560, -33.31052, -34.52192, -34.52192),
  Lon = c(152.49460, 151.48528, 150.84102, 150.84102),
  Estuary = c("Wallis Lake", "Tuggerah Lake", "Lake Illawarra", "Lake Illawarra")
)

#NOT RUN
#result <- download_process_wind_data(
 # locations = sample_locations,  # Required dataframe with Lat, Lon, and location names
 # start_year = 2011,             # Start year for data (default: 2009)
 # end_year = 2012,               # End year for data (default: 2024)
 # output_dir = "Wind_Data",      # Directory for output files
 # save_interim = TRUE,           # Save intermediate files
 # download_u = TRUE,             # Download U wind data
 # download_v = TRUE,             # Download V wind data
 # overwrite_existing = T     # Skip existing files (default: FALSE)
#)
}
