% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_flow_data.R
\name{get_water_flow_data}
\alias{get_water_flow_data}
\title{Download WaterNSW Gauge Flow Data}
\usage{
get_water_flow_data(sites, date_start, date_end, token)
}
\arguments{
\item{sites}{a vector of WaterNSW site codes (as strings)}

\item{date_start}{a date in format 'DD-MMM-YYYY HH:MM' (as string)}

\item{date_end}{a date in format 'DD-MMM-YYYY HH:MM' (as string)}

\item{token}{the Water NSW API Subscription Key}
}
\value{
Dataframe containing daily flow data
}
\description{
This function pulls flow data from the NSW Water Online API. https://api-portal.waternsw.com.au/reports.
Specify sites by number and a date range.
You need to have a valid API token to use the function.
There are rather restrictive limits on the API basic tier (28 requests per 24 hours and a limit of 1000 rows of data per "request")
Large data requests therefore use more API tokens.
}
\examples{
# Example Usage
token = keyring::key_get("NSW_Water_Online") # this pulls a saved key from the computer,
# set with key_set("NSW_Water_Online")
sites <- c("204007","207009","2122001")
date_start <- "01-Jan-2022 00:00"
date_end <- "01-Jan-2024 00:00"

#NOT RUN
# water_data <- get_water_flow_data(sites = sites,
# date_start = date_start, date_end = date_end, token = water_token)
}
